% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctracker.R
\name{BuildCCDataFrame}
\alias{BuildCCDataFrame}
\title{Read raw data and returns as data.frame}
\usage{
BuildCCDataFrame(local.file = paste(tempdir(), "cctracker.txt", sep =
  ifelse(.Platform$OS.type == "windows", "\\\\", "/")))
}
\arguments{
\item{local.file}{}
}
\value{
data.frame
}
\description{
Read raw data and returns as data.frame
}
