% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehol.R
\name{BuildFHODataFrame}
\alias{BuildFHODataFrame}
\title{Read raw data and returns as data.frame}
\usage{
BuildFHODataFrame(local.file = paste(tempdir(), "firehol.txt", sep =
  ifelse(.Platform$OS.type == "windows", "\\\\", "/")))
}
\arguments{
\item{local.file}{}
}
\value{
data.frame
}
\description{
Read raw data and returns as data.frame
}
