% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailonetime.R
\name{BuildMOTDataFrame}
\alias{BuildMOTDataFrame}
\title{Read raw data and returns as data.frame}
\usage{
BuildMOTDataFrame(local.file = paste(tempdir(), "mailonetime.txt", sep =
  ifelse(.Platform$OS.type == "windows", "\\\\", "/")))
}
\arguments{
\item{local.file}{}
}
\value{
data.frame
}
\description{
Read raw data and returns as data.frame
}
