% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwr.domains.R
\name{BuildMWDDataFrame}
\alias{BuildMWDDataFrame}
\title{Read raw data and returns as data.frame}
\usage{
BuildMWDDataFrame(local.file = paste(tempdir(), "malware.domains.txt", sep =
  ifelse(.Platform$OS.type == "windows", "\\\\", "/")))
}
\arguments{
\item{local.file}{}
}
\value{
data.frame
}
\description{
Read raw data and returns as data.frame
}
