% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishtank.R
\name{BuildPTDDataFrame}
\alias{BuildPTDDataFrame}
\title{Read raw data and returns as data.frame}
\usage{
BuildPTDDataFrame(local.file = paste(tempdir(), "phishtank.txt", sep =
  ifelse(.Platform$OS.type == "windows", "\\\\", "/")))
}
\arguments{
\item{local.file}{}
}
\value{
data.frame
}
\description{
Read raw data and returns as data.frame
}
