% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransom.tracker.R
\name{BuildRWDDataFrame}
\alias{BuildRWDDataFrame}
\title{Read raw data and returns as data.frame}
\usage{
BuildRWDDataFrame(local.file = paste(tempdir(), "ransom.tracker.domains.txt",
  sep = ifelse(.Platform$OS.type == "windows", "\\\\", "/")))
}
\arguments{
\item{local.file}{}
}
\value{
data.frame
}
\description{
Read raw data and returns as data.frame
}
