% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransom.tracker.R
\name{BuildRWIDataFrame}
\alias{BuildRWIDataFrame}
\title{Read raw data and returns as data.frame}
\usage{
BuildRWIDataFrame(local.file = paste(tempdir(), "ransom.tracker.ips.txt", sep
  = ifelse(.Platform$OS.type == "windows", "\\\\", "/")))
}
\arguments{
\item{local.file}{}
}
\value{
data.frame
}
\description{
Read raw data and returns as data.frame
}
