% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransom.tracker.R
\name{BuildRWUDataFrame}
\alias{BuildRWUDataFrame}
\title{Read raw data and returns as data.frame}
\usage{
BuildRWUDataFrame(local.file = paste(tempdir(), "ransom.tracker.urls.txt", sep
  = ifelse(.Platform$OS.type == "windows", "\\\\", "/")))
}
\arguments{
\item{local.file}{}
}
\value{
data.frame
}
\description{
Read raw data and returns as data.frame
}
