% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornodes.R
\name{BuildTorDataFrame}
\alias{BuildTorDataFrame}
\title{Read raw data and returns as data.frame}
\usage{
BuildTorDataFrame(local.file = paste(tempdir(), "tornodes.txt", sep =
  ifelse(.Platform$OS.type == "windows", "\\\\", "/")))
}
\arguments{
\item{local.file}{}
}
\value{
data.frame
}
\description{
Read raw data and returns as data.frame
}
