% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inet.intel.R
\name{CheckIOC}
\alias{CheckIOC}
\title{Check if IOC is in data.frame. It returns info as json.}
\usage{
CheckIOC(ioc = "github.com")
}
\arguments{
\item{ioc}{}
}
\value{
json
}
\description{
Check if IOC is in data.frame. It returns info as json.
}
