% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firehol.R
\name{DownloadFHOData}
\alias{DownloadFHOData}
\title{CIDR Black list
https://raw.githubusercontent.com/firehol/blocklist-ipsets/master/firehol_level1.netset}
\usage{
DownloadFHOData(dstpath = tempdir())
}
\arguments{
\item{dstpath}{path where source data will be stored, tempdir set as default.}
}
\value{
character, local path of downloaded file
}
\description{
Download source data to local file
}
