% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailonetime.R
\name{DownloadMOTData}
\alias{DownloadMOTData}
\title{One time email domains
Ref: https://raw.githubusercontent.com/martenson/disposable-email-domains/master/disposable_email_blacklist.conf
MOT = Mail One Time
Download source data to local file}
\usage{
DownloadMOTData(dstpath = tempdir())
}
\arguments{
\item{dstpath}{path where source data will be stored, tempdir set as default.}
}
\value{
character, local path of downloaded file
}
\description{
One time email domains
Ref: https://raw.githubusercontent.com/martenson/disposable-email-domains/master/disposable_email_blacklist.conf
MOT = Mail One Time
Download source data to local file
}
