% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlwr.domains.R
\name{DownloadMWDData}
\alias{DownloadMWDData}
\title{Malware domains
Ref: http://mirror1.malwaredomains.com/files/
http://mirror1.malwaredomains.com/files/justdomains
Download source data to local file}
\usage{
DownloadMWDData(dstpath = tempdir())
}
\arguments{
\item{dstpath}{path where source data will be stored, tempdir set as default.}
}
\value{
character, local path of downloaded file
}
\description{
Malware domains
Ref: http://mirror1.malwaredomains.com/files/
http://mirror1.malwaredomains.com/files/justdomains
Download source data to local file
}
