% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phishtank.R
\name{DownloadPTDData}
\alias{DownloadPTDData}
\title{PhisTank
Ref: https://www.phishtank.com/developer_info.php
http://data.phishtank.com/data/online-valid.csv
Download source data to local file}
\usage{
DownloadPTDData(dstpath = tempdir())
}
\arguments{
\item{dstpath}{path where source data will be stored, tempdir set as default.}
}
\value{
character, local path of downloaded file
}
\description{
PhisTank
Ref: https://www.phishtank.com/developer_info.php
http://data.phishtank.com/data/online-valid.csv
Download source data to local file
}
