% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransom.tracker.R
\name{DownloadRWDData}
\alias{DownloadRWDData}
\title{Ransomware Tracker
Ref: https://ransomwaretracker.abuse.ch/blocklist/
Domain: https://ransomwaretracker.abuse.ch/downloads/RW_DOMBL.txt
URL: https://ransomwaretracker.abuse.ch/downloads/RW_URLBL.txt
IP: https://ransomwaretracker.abuse.ch/downloads/RW_IPBL.txt}
\usage{
DownloadRWDData(dstpath = tempdir())
}
\arguments{
\item{dstpath}{path where source data will be stored, tempdir set as default.}
}
\value{
character, local path of downloaded file
}
\description{
Download source data to local file
}
