% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransom.tracker.R
\name{DownloadRWUData}
\alias{DownloadRWUData}
\title{Download source data to local file}
\usage{
DownloadRWUData(dstpath = tempdir())
}
\arguments{
\item{dstpath}{path where source data will be stored, tempdir set as default.}
}
\value{
character, local path of downloaded file
}
\description{
Download source data to local file
}
