% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornodes.R
\name{DownloadTorData}
\alias{DownloadTorData}
\title{Tor nodes data from Daniel (me@dan.me.uk)
No more than 1 query every 30min
Ref: https://www.dan.me.uk/tornodes
Download source data to local file}
\usage{
DownloadTorData(dstpath = tempdir())
}
\arguments{
\item{dstpath}{path where source data will be stored, tempdir set as default.}
}
\value{
character, local path of downloaded file
}
\description{
Tor nodes data from Daniel (me@dan.me.uk)
No more than 1 query every 30min
Ref: https://www.dan.me.uk/tornodes
Download source data to local file
}
